% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.wordcloud}
\alias{radiant.wordcloud}
\alias{radiant.wordcloud-package}
\title{radiant.wordcloud}
\usage{
radiant.wordcloud()
}
\description{
radiant.wordcloud

Launch Radiant in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}

