% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{wordcloud}
\alias{wordcloud}
\title{wordcloud}
\format{A data frame with 18 rows and 5 variables}
\usage{
data(wordcloud)
}
\description{
wordcloud
}
\details{
Rankings reflect the evaluation of 18 alternative carpet cleaners by one respondent. Description provided in attr(carpet,"description")
}
\keyword{datasets}

